app/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
appnope==0.1.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.6
comm==0.1.3
debugpy==1.6.7
decorator==5.1.1
exceptiongroup==1.1.2
executing==1.2.0
Flask==2.3.2
frozenlist==1.4.0
h11==0.14.0
idna==3.4
ipykernel==6.24.0
ipython==8.14.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.3.0
jupyter_core==5.3.1
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
multidict==6.0.4
nest-asyncio==1.5.6
numpy==1.25.2
oauthlib==3.2.2
openai==0.27.8
outcome==1.2.0
packaging==23.1
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.8.1
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pyzmq==25.1.0
requests==2.31.0
requests-oauthlib==1.3.1
selenium==4.11.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
stack-data==0.6.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
trio==0.22.2
trio-websocket==0.10.3
tweepy==4.14.0
tzdata==2023.3
urllib3==2.0.3
wcwidth==0.2.6
Werkzeug==2.3.6
wsproto==1.2.0
yarl==1.9.2
